wifi: Avoiding kernel crash with os_SignalObjectWait timeout
BZ: 5868

In case os_SignalObjectWait does timeout there are possibilities of crash.

After os_SignalObjectWait timeout driver frees the command memory while the

command is still on the queue. So next time when the command is about to be
executed que_Dequeue will try to access the memory which is already freed and
causes a crash.

This change is for avoiding this crash

Change-Id: I249d0f89ecc526198aa63a393a3b10e6d4e2cbdf
Signed-off-by: Jeremie GARCIA <jeremiex.garcia@intel.com>