wifi: boot_without_wifi_chip
BZ: 6415

In condition that no wifi chip has ever been plugged in, the board
freezes at boot time.
This is implicitely issued by the calibration method, that look for a
nvs file at boot time. If not found, it performs the calibration in which
we try to deal with the driver. However, no chip is there and it leads to
such freezes.

At MCP driver init (on insmod), we try to detect the chip. When not detected,
the insmod init method returns an error and calibration stops there.

We are now able to boot the board without wifi chip.
We are able as well to plugg a chip and reboot to have both calibration and
wifi feature enabled.

Change-Id: Ib1fd73e324fdcce10019c112c00ce32b9e0ff188
Signed-off-by: Jeremie GARCIA <jeremiex.garcia@intel.com>