[PORT FROM R2] widi: EAPOL start race condtion between assoc_done and first EAP packet
BZ: 17282

we might get sometimes the first EAPOL from the client
before we got the tx result of the association response (and thus the station
wasn't added yet).
in this case, mac80211 will drop the packet, as it's a data packet from unknown
station.
then, the GO doesn't get the EAPOL Start from the device, and the connection
fails.

This patch fixes this race condition in MAC80211.It has been provided by TI and
is mapped on CSR OMAPS00256120

Signed-off-by: Jean-Michel Bachot <jean-michelx.bachot@intel.com>

Change-Id: Ia34ff2ea59a00c0d2dc2cf7a771a4f5d62aa3e3b
Orig-Change-Id: If2f3498f5409d87ed4e5a96dd65e0fa61ab89e5a
Signed-off-by: Jean-Michel Bachot <jean-michelx.bachot@intel.com>
Reviewed-on: http://android.intel.com:8080/27204
Reviewed-by: Trivelly, JeanX <jeanx.trivelly@intel.com>
Reviewed-by: Regairaz, JulieX <juliex.regairaz@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Signed-off-by: Jean-Michel Bachot <jean-michelx.bachot@intel.com>
Reviewed-on: http://android.intel.com:8080/31889
Reviewed-by: Marotte, Fabien <fabien.marotte@intel.com>