[PORT FROM R2] wl12xx: Avoid toggling WLAN_EN on S3 when the WIFI is disabled.
BZ: 18765 17650

When the driver is unloaded from the kernel, the mmc driver will
not be able to handle the power card properly on suspend/resume.

By enabling the manual resume of the mmc driver when wl12xx driver
is unloaded, the mmc driver will not control the power of the sdio
card.

By the same way, by disabling the manual resume of the mmc driver
when the wl12xx in loaded, the mmc driver will be able to control
the power supply of the sdio card depending on pm_flags (if
MMC_PM_KEEP_POWER is set or not).

Using this technique avoid toggling ON/OFF the wl12xx when the
WIFI is disabled during suspend / resume transitions.

Note that the sdio slot from the host side shall initialise
bus_resume_flags to MMC_BUSRESUME_MANUAL_RESUME.

Signed-off-by: Aymen Zayet <aymenx.zayet@intel.com>
Orig-Change-Id: I30b6389290f4fc13115c06c67fd2538d21aea230
Change-Id: I30b6389290f4fc13115c06c67fd2538d21aea230
Reviewed-on: http://android.intel.com:8080/34512
Reviewed-by: Garcia, JeremieX <jeremiex.garcia@intel.com>
Reviewed-by: Regairaz, JulieX <juliex.regairaz@intel.com>
Reviewed-by: Fiat, Christophe <christophe.fiat@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>