[PORT FROM R2] WLAN: Disconnection reason 4 = disconnection due to inactivity.
BZ: 21938 17650

Disconnection is due to AP not acking NullData packet
with Power Save Mode flag set. It is potentially due to
AP doing some other tasks else (this is AP proprietary algo).

Force sending NullData in Active Mode in order to unlock
the situation and avoid dropping the connection in this case.

Change-Id: Ida17ecbee21b47ef4c234e792ed5b12f4e0c8cd7
Orig-Change-Id: I41fe76ebe519d7cc84b18f97c38c72df2a81b895
Signed-off-by: Jean Trivelly <jeanx.trivelly@intel.com>
Reviewed-on: http://android.intel.com:8080/35271
Reviewed-by: Garcia, JeremieX <jeremiex.garcia@intel.com>
Reviewed-by: Regairaz, JulieX <juliex.regairaz@intel.com>
Reviewed-by: Bachot, Jean-michelX <jean-michelx.bachot@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>