wlan: enable/disable manual resume on power on and power off
BZ: 32339

Now that wlan driver calls probe on boot and never
calls remove, the manual resume was left disabled
even when wifi was not turned on. So that the enabling
and disabling of manual suspend follows wlan enable/disable
it is better done on the turn on and off functions.

Change-Id: I739f36efdae92a8d128f274ea2280be0cc3b5029
Signed-off-by: Axel Haslam <axelx.haslam@intel.com>
Reviewed-on: http://android.intel.com:8080/40025
Reviewed-by: Zayet, AymenX <aymenx.zayet@intel.com>
Reviewed-by: Regairaz, JulieX <juliex.regairaz@intel.com>
Reviewed-by: De Cesco, Jonathan <jonathan.de.cesco@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>