wlan_prov: Make the provisionning platform independant.
BZ: 20626

To bind / unbind the device from the driver, pci ID is needed.
This one may change depending on the firmware. So to keep this
application independant from the platform, the device ID of the
SDIO controller attached to the wifi card is detected dynamically
by looking at /sys/bus/sdio/devices/.

Change-Id: I3147ca740eb3adf42dc4b9dc15f7e1aa6ac24c24
Signed-off-by: Aymen Zayet <aymenx.zayet@intel.com>
Reviewed-on: http://android.intel.com:8080/41629
Reviewed-by: De Cesco, Jonathan <jonathan.de.cesco@intel.com>
Reviewed-by: Tricot, HonoreX <honorex.tricot@intel.com>
Reviewed-by: Trivelly, JeanX <jeanx.trivelly@intel.com>
Reviewed-by: Regairaz, JulieX <juliex.regairaz@intel.com>
Reviewed-by: Bachot, Jean-michelX <jean-michelx.bachot@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>