wifi: avoid wiphy_unregister panic
BZ: 32407

When a memory allocation fails in mac80211 if_add registration path,
the kernel is trying to rollback this registration and ends up in a
panic:
[] BUG: unable to handle kernel NULL pointer dereference at 00000004
[] IP: [<fa0d223c>] wiphy_unregister+0x31c/0x670 [cfg80211]
This is due to unregister netdev events being called, when
the pendant register events were never called. This lead to trying to
free a list that was never initialized. Workaround this by insuring
the overall netdev_list is also not empty when removing the wireless
device from the list
This patch should only be merged if issue is reproduced after insmod
/rmmod patches are integrated

Change-Id: I0be254017cc984c72c6897466363cabca5027df5
Signed-off-by: Jonathan DE CESCO <jonathan.de.cesco@intel.com>
Reviewed-on: http://android.intel.com:8080/42943
Reviewed-by: Regairaz, JulieX <juliex.regairaz@intel.com>
Reviewed-by: Zayet, AymenX <aymenx.zayet@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>