WLAN: DUT wait about 30 seconds to connect to hidden AP.
BZ: 28985

This is a fix which set the minimum and the maximum dwell time to the
same value (30ms). Indeed, some AP take to long to respond to probe
request and are therefore not taken into account by the STA.
The dwell time is now fixed and no longer dynamic.

Change-Id: I68dd08bb6fb4eb4d1d16cfe0ec861708dc1dfe4f
Signed-off-by: Jean-Marie Lemetayer <jean-mariex.lemetayer@intel.com>
Reviewed-on: http://android.intel.com:8080/44149
Reviewed-by: De Cesco, Jonathan <jonathan.de.cesco@intel.com>
Reviewed-by: Regairaz, JulieX <juliex.regairaz@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>