wl12xx: Reset channels configuration on wl1271_init_ieee80211()
BZ: 31993

The Calibrator configures the channels through netlink commands based
on the default NVS firmware. After this process, the new NVS is loaded
but wl1271_channels[] and wl1271_channels_5ghz[] arrays used by mac80211
remain corrupted by the calibration process. As a result, regulatory
domains are not consistent for 2 and 5 Ghz bands.

To resolve this issue, two const arrays are defined to keep the default
configuration for both bands and the configurations of the channels
supported are copied from their corresponding default values each time
wl1271_init_ieee80211() is called.

Change-Id: Icdef4739d29ec838d945f293fa6af756d0c20f3c
Signed-off-by: Aymen Zayet <aymenx.zayet@intel.com>
Reviewed-on: http://android.intel.com:8080/44341
Reviewed-by: Regairaz, JulieX <juliex.regairaz@intel.com>
Reviewed-by: Garcia, JeremieX <jeremiex.garcia@intel.com>
Reviewed-by: Trivelly, JeanX <jeanx.trivelly@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: De Cesco, Jonathan <jonathan.de.cesco@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>