[PORT FROM R2] wifi: increase ps_poll_timeouts
BZ: 30112

Increase the time the device waits for data after a PS-Poll or
UAPSD packet.
For "normal" behavior, this has no impact as this timer expires
before the timoeut as soon as data is received.
For noisy environments and/or bad AP behavior, the device will
stay active more longer, and is more likely to sustain the
connection and throughput

Orig-Change-Id: I84be05e9e274cc6b1d6574e6c374e426d4e4b714
Change-Id: Ia85472743a7a5790fdb07e4c07c1b0aaaeaa5239
Signed-off-by: Jonathan DE CESCO <jonathan.de.cesco@intel.com>
Reviewed-on: http://android.intel.com:8080/46015
Reviewed-by: Isaksen, FrodeX <frodex.isaksen@intel.com>
Reviewed-by: Lemetayer, Jean-marieX <jean-mariex.lemetayer@intel.com>
Reviewed-by: Trivelly, Jean <jean.trivelly@intel.com>
Reviewed-by: Regairaz, Julie <julie.regairaz@intel.com>
Reviewed-by: Bachot, Jean-michel <jean-michel.bachot@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>