wifi: prevent list corruption in unlink_bss
BZ: 34090

This patch partially reverts
"cfg80211: more correct support of multi-essid bss-es"
which lead to regression/panic with some APs in hidden mode

Patch provided by TI to fix a kernel panic due to slab corruption

TI CSR OMAPS00261359

Looks like bss_release deallocates memory with wrong
memory pointer. However, it's not possible to find
how it had wrong pointer in the structure. There's
a patch that is handling the structure. The patch
is not for fixing exact same problem, but it includes
code that handles related structure.

Change-Id: I562c658cf07474cf79e88feb9b8372e9831a94d1
Signed-off-by: Jean-Michel Bachot <jean-michel.bachot@intel.com>