WLAN: fix provisioning issue when change MAC address without calibration
BZ: 49210

unbind/bind sequence was move just after calibration to avoid unnecessary
unbind/bind that sometime cause failure. But one case was forget, when
only updating MAC address without previous calibration. In this case,
the unbind / bind of the driver is no more done. So the the new nvs file
is not taking into accout till rebooting the device.

Fix this issue by moving back the unbind/bind sequence at the end of
provisioning, but enable a trigger to start it only when calibration is
done and/or MAC updated.

Change-Id: I7c4d29a1c06d89fe658b528382bc0f1b58107aac
Orig-Change-Id: I735e402d5b31fe49b34972fb2d3c8ef31f65fb32
Signed-off-by: Honore Tricot <honorex.tricot@intel.com>
Signed-off-by: jonathan.de.cesco <jonathan.de.cesco@intel.com>
Reviewed-on: http://android.intel.com:8080/61810
Reviewed-by: Regairaz, Julie <julie.regairaz@intel.com>
Reviewed-by: Lemetayer, Jean-marieX <jean-mariex.lemetayer@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Yin, Fengwei <fengwei.yin@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/77971
Reviewed-by: Li, XiaojingX <xiaojingx.li@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>