wl12xx: release wake lock when removing the interface.
BZ: 48052

When the hard IRQ is called, a wake lock is held and the thread irq
is scheduled later time. This one is supposed to release the wakelock.
If a recovery happens in between, the interrupt are disabled without
irq_synchronize(). As a consequence, the thread irq will be cancelled
and the wake lock remains held.

Moreover, recovery calls __wl1271_op_remove_interface() which clears
wl->flags, making the call to wake_lock() / wake_unlock() unbalanced.

To fix this issue, WL1271_FLAG_WAKE_LOCK flag should be checked on
wl1271_op_remove_interface() and the wake_lock should be called
accordingly.

Change-Id: Ie7657ec3ba67013b27e1df69d25ee35e86762c87
Orig-Change-Id: I520daf004339ca7339b31de60be066981ba59a5e
Signed-off-by: Aymen Zayet <aymen.zayet@intel.com>
Signed-off-by: Axel Haslam <axelx.haslam@intel.com>
Reviewed-on: http://android.intel.com:8080/62660
Reviewed-by: Regairaz, Julie <julie.regairaz@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/77976
Reviewed-by: Li, XiaojingX <xiaojingx.li@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>