wifi: failed to scan if recovery during scan happens
BZ: 56763

mac80211 layer can be stuck in scan busy state if a scan
command is issued while recovery is in progress but driver
state is not off. Fix it by rejecting scan command if this state is detected.
More generally, add more robustness to avoid being locked in this scan busy
state, and fix it also at scan cancel operation.

Orig-Change-Id: Ie02da408e846ad3e3fd9cc4c3a20f359cf62750a
Change-Id: If99bd199f88ec622aea458998080377a3bcd588c
Signed-off-by: Jonathan DE CESCO <jonathan.de.cesco@intel.com>
Reviewed-on: http://android.intel.com:8080/68319
Reviewed-by: Trivelly, Jean <jean.trivelly@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/77989
Reviewed-by: Li, XiaojingX <xiaojingx.li@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>