wl12xx: add support for U-APSD powersave for Voice traffic
BZ: 43272

Add support for U-APSD powersave by forcing the
chip in power save mode if N (=10) consecutive packets
transmitted on the AC_VO queue. The time to go into
power save with RTP traffic should then be about 200ms.
This should avoid going in and out of power save if a
mix of Voice traffic with "normal" traffic is sent.
The forced power save mode is left if a "normal"
packet is sent.

Signed-off-by: Frode Isaksen <fisaksex>

Change-Id: Ib7d1c8851553aad8bd0a5ce792fed7e6d0713682
Reviewed-on: http://android.intel.com:8080/69635
Reviewed-by: Isaksen, FrodeX <frodex.isaksen@intel.com>
Reviewed-by: Fiat, Christophe <christophe.fiat@intel.com>
Reviewed-by: De Cesco, Jonathan <jonathan.de.cesco@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/77996
Reviewed-by: Li, XiaojingX <xiaojingx.li@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>