wl12xx: Always log the packets that caused WiFi to wake up the platform
BZ: 62772

Finding why WiFi woke up the platform can be difficult involving using Wireshark or
Omnipeek network sniffers.
This patch will log the N packets (N=2) that is received after the WiFi chip
woke up the platform. With N=2, we will hopefully log the packet that
caused the wakeup, since typically we have an ARP packet and an IP packet.
The changes from patch 64576 is that the log is now printed as KERN_INFO as
to always have the log.

Change-Id: Icbef3d2cfd69cfa14323111f43ceddc43c27308d
Signed-off-by: Frode Isaksen <fisaksex>
Reviewed-on: http://android.intel.com:8080/70609
Reviewed-by: Isaksen, FrodeX <frodex.isaksen@intel.com>
Reviewed-by: De Cesco, Jonathan <jonathan.de.cesco@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/77997
Reviewed-by: Li, XiaojingX <xiaojingx.li@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>