[PORT FROM JBMR0]wl12xx: remove useless wakelock
BZ: 70447

wl->wake_lock is used to avoid S3 between the hardirq and the thread
irq handler. This change is not needed anymore since interrupts are
synchronized on suspend.

NB: This change is mandatory on the kernel 3.4 because wakesource are
considered as a wake event. So even if the wakelock is released during
suspend, the system  will reject S3 and wakes up.

Change-Id: I08ea7f8d4aa0ebba6cdd2f345023c68931a6cd6a
Orig-Change-Id: I599978048b594df35277d2485c2fc95493ede90c
Signed-off-by: Aymen Zayet <aymen.zayet@intel.com>
Signed-off-by: Axel Haslam <axelx.haslam@intel.com>
Reviewed-on: http://android.intel.com:8080/85622
Reviewed-by: Fiat, Christophe <christophe.fiat@intel.com>
Reviewed-by: Casasnovas, QuentinX <quentinx.casasnovas@intel.com>
Reviewed-by: De Cesco, Jonathan <jonathan.de.cesco@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: Kuppuswamy, Sathyanarayanan <sathyanarayanan.kuppuswamy@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>