wl12xx: Fix possible panic in resume path
BZ: 75240

During resume, it is possible that there was a pending work when we suspended
the chip. In this case it is important to do that work (i.e. running wl12xx_irq_locked())
but it may happen that this work trigger a firmware recovery. In this case,
it is very important to leave the function with an error and not try to resume
the chip anyway (or send any command to the chip) as the firmware is already
crashed.

This patch makes sure we leave the function if treating the pending work triggers a
recovery and also removes some dead code.

Change-Id: I5a815ad5d7a15a108930f50749f978dd6dcaa6da
Signed-off-by: Quentin Casasnovas <quentinx.casasnovas@intel.com>
Reviewed-on: http://android.intel.com:8080/85337
Reviewed-by: De Cesco, Jonathan <jonathan.de.cesco@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>